tests/submission/PhilipTaylor/canvas/2d.state.saverestore.underflow.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: 2d.state.saverestore.underflow</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.state.html">state</a>.<a href="index.2d.state.saverestore.html">saverestore</a>.underflow</h1>
<p class="desc">restore() with an empty stack has no effect</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.2d.state.restore.underflow">2d.state.restore.underflow</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("restore() with an empty stack has no effect");
_addTest(function(canvas, ctx) {

for (var i = 0; i < 16; ++i)
    ctx.restore();
ctx.lineWidth = 0.5;
ctx.restore();
_assertEqual(ctx.lineWidth, 0.5, "ctx.lineWidth", "0.5");


});
</script>

<div id="log"></div>
</body></html>
